+++ terraform/dynamodb/dynamodb.tf
@@ -1,0 +1,14 @@
+# dynamodb/dynamodb.tf
+
+resource "aws_dynamodb_table" "example" {
+  name           = "my-table"
+  billing_mode   = "PROVISIONED"
+  read_capacity  = 20
+  write_capacity = 20
+  hash_key       = "id"
+
+  attribute {
+    name = "id"
+    type = "S"
+  }
+}