+++ terraform/lambda/variables.tf
@@ -1,0 +1,6 @@
+# lambda/variables.tf
+
+variable "dynamodb_table_name" {
+  description = "The name of the DynamoDB table"
+  type        = string
+}